--- script/prepublish
@@ -1,5 +1,0 @@
-#!/bin/bash -e
-
-echo "Trying to publish..."
-
-gulp all
