+++ script/publish
@@ -8,5 +8,6 @@ fi
 
 echo "Trying to publish..."
 
+gulp all
 script/version-bump
 npm publish
